Simplify gokeyless-tester by moving common test elements to the tests package.
