+++ cmd/gokeyless-tester/gokeyless-tester.go
@@ -1,18 +1,11 @@
 package main
 
 import (
-	"bytes"
 	"crypto"
-	"crypto/ecdsa"
-	"crypto/rand"
-	"crypto/rsa"
 	"crypto/x509"
-	"encoding/asn1"
 	"encoding/pem"
-	"errors"
 	"flag"
 	"io/ioutil"
-	"math/big"
 	"os"
 	"path/filepath"
 	"regexp"
@@ -21,6 +14,7 @@ import (
 	"github.com/cloudflare/cfssl/helpers"
 	"github.com/cloudflare/cfssl/log"
 	"github.com/cloudflare/gokeyless/client"
+	"github.com/cloudflare/gokeyless/tests"
 )
 
 var (
@@ -83,48 +77,7 @@ func main() {
 		}
 	}
 
-	log.Infof("Testing %s for %v with %d workers...", server, testLen, workers)
-	errCount := 0
-
-	errs := loadTest(func() error {
-		if err := testConnect(c, server); err != nil {
-			return err
-		}
-
-		for _, key := range privkeys {
-			if err := testKey(key); err != nil {
-				return err
-			}
-		}
-		return nil
-	})
-
-	done := time.After(testLen)
-	for {
-		select {
-		case err := <-errs:
-			log.Error(err)
-			errCount++
-
-		case <-done:
-			log.Infof("Completed with %d errors", errCount)
-			return
-		}
-	}
-}
-
-func loadTest(test func() error) <-chan error {
-	errs := make(chan error)
-	for i := 0; i < workers; i++ {
-		go func() {
-			for {
-				if err := test(); err != nil {
-					errs <- err
-				}
-			}
-		}()
-	}
-	return errs
+	tests.RunServerTests(testLen, workers, c, server, privkeys)
 }
 
 // LoadPubKey attempts to load a public key from PEM or DER.
@@ -187,88 +140,4 @@ func LoadCertsFromDir(dir string) (certs []*x509.Certificate, err error) {
 		return nil
 	})
 	return
-}
-
-func testConnect(c *client.Client, server string) error {
-	conn, err := c.Dial(server)
-	if err != nil {
-		return err
-	}
-	defer conn.Close()
-	return conn.Ping(nil)
-}
-
-func hashPtxt(h crypto.Hash, ptxt []byte) []byte {
-	return h.New().Sum(ptxt)[len(ptxt):]
-}
-
-func testKey(key *client.PrivateKey) (err error) {
-	ptxt := []byte("Hello!")
-	r := rand.Reader
-	hashes := []crypto.Hash{
-		crypto.MD5SHA1,
-		crypto.SHA1,
-		crypto.SHA224,
-		crypto.SHA256,
-		crypto.SHA384,
-		crypto.SHA512,
-	}
-
-	for _, h := range hashes {
-		var msg, sig []byte
-		if h == crypto.MD5SHA1 {
-			msg = append(hashPtxt(crypto.MD5, ptxt), hashPtxt(crypto.SHA1, ptxt)...)
-		} else {
-			msg = hashPtxt(h, ptxt)
-		}
-
-		if sig, err = key.Sign(r, msg, h); err != nil {
-			return
-		}
-
-		switch pub := key.Public().(type) {
-		case *rsa.PublicKey:
-			if err = rsa.VerifyPKCS1v15(pub, h, msg, sig); err != nil {
-				return
-			}
-		case *ecdsa.PublicKey:
-			ecdsaSig := new(struct{ R, S *big.Int })
-			asn1.Unmarshal(sig, ecdsaSig)
-			if !ecdsa.Verify(pub, msg, ecdsaSig.R, ecdsaSig.S) {
-				return errors.New("ecdsa verify failed")
-			}
-		default:
-			return errors.New("unknown public key type")
-		}
-	}
-
-	if pub, ok := key.Public().(*rsa.PublicKey); ok {
-		var c, m []byte
-		if c, err = rsa.EncryptPKCS1v15(r, pub, ptxt); err != nil {
-			return
-		}
-
-		if m, err = key.Decrypt(r, c, &rsa.PKCS1v15DecryptOptions{}); err != nil {
-			return
-		}
-		if bytes.Compare(ptxt, m) != 0 {
-			return errors.New("rsa decrypt failed")
-		}
-
-		if m, err = key.Decrypt(r, c, &rsa.PKCS1v15DecryptOptions{SessionKeyLen: len(ptxt)}); err != nil {
-			return
-		}
-		if bytes.Compare(ptxt, m) != 0 {
-			return errors.New("rsa decrypt failed")
-		}
-
-		if m, err = key.Decrypt(r, c, &rsa.PKCS1v15DecryptOptions{SessionKeyLen: len(ptxt) + 1}); err != nil {
-			return
-		}
-		if bytes.Compare(ptxt, m) == 0 {
-			return errors.New("rsa decrypt suceeded despite incorrect SessionKeyLen")
-		}
-	}
-
-	return nil
 }
