+++ tests/tests.go
@@ -1,2 +1,160 @@
-// Package tests exists just as a common place for all gokeyless tests.
 package tests
+
+import (
+	"bytes"
+	"crypto"
+	"crypto/ecdsa"
+	"crypto/rand"
+	"crypto/rsa"
+	"encoding/asn1"
+	"errors"
+	"math/big"
+	"time"
+
+	"github.com/cloudflare/cfssl/log"
+	"github.com/cloudflare/gokeyless/client"
+)
+
+//RunServerTests load tests a server for a given ammount of time and
+func RunServerTests(testLen time.Duration, workers int, c *client.Client, server string, privkeys []crypto.Signer) {
+	log.Infof("Testing %s for %v with %d workers...", server, testLen, workers)
+	errCount := 0
+
+	running := make(chan bool, workers)
+	errs := spawnTesters(workers, running, func() error {
+		if err := testConnect(c, server); err != nil {
+			return err
+		}
+
+		for _, priv := range privkeys {
+			if err := testKey(priv); err != nil {
+				return err
+			}
+		}
+		return nil
+	})
+
+	timeout := time.After(testLen)
+	for {
+		select {
+		case err := <-errs:
+			running <- true
+			if err != nil {
+				log.Error(err)
+				errCount++
+			}
+
+		case <-timeout:
+			close(running)
+			log.Infof("Completed with %d errors", errCount)
+			return
+		}
+	}
+}
+
+// spawnTesters spawns a number of goroutines to repeatedly perform a test.
+// Any errors are returned on a channel
+func spawnTesters(numWorkers int, running chan bool, test func() error) <-chan error {
+	errs := make(chan error)
+	for i := 0; i < numWorkers; i++ {
+		go func() {
+			running <- true
+			for range running {
+				errs <- test()
+			}
+		}()
+	}
+	return errs
+}
+
+// testConnect tests the ability to Dial and Ping a given server
+func testConnect(c *client.Client, server string) error {
+	conn, err := c.Dial(server)
+	if err != nil {
+		return err
+	}
+	defer conn.Close()
+	return conn.Ping(nil)
+}
+
+// hashPtxt hashes the plaintext with the given hash algorithm.
+func hashPtxt(h crypto.Hash, ptxt []byte) []byte {
+	return h.New().Sum(ptxt)[len(ptxt):]
+}
+
+// testKey performs and verifies all possible opaque private key operations.
+func testKey(priv crypto.Signer) (err error) {
+	ptxt := []byte("Hello!")
+	r := rand.Reader
+	hashes := []crypto.Hash{
+		crypto.MD5SHA1,
+		crypto.SHA1,
+		crypto.SHA224,
+		crypto.SHA256,
+		crypto.SHA384,
+		crypto.SHA512,
+	}
+
+	for _, h := range hashes {
+		var msg, sig []byte
+		if h == crypto.MD5SHA1 {
+			msg = append(hashPtxt(crypto.MD5, ptxt), hashPtxt(crypto.SHA1, ptxt)...)
+		} else {
+			msg = hashPtxt(h, ptxt)
+		}
+
+		if sig, err = priv.Sign(r, msg, h); err != nil {
+			return
+		}
+
+		switch pub := priv.Public().(type) {
+		case *rsa.PublicKey:
+			if err = rsa.VerifyPKCS1v15(pub, h, msg, sig); err != nil {
+				return
+			}
+		case *ecdsa.PublicKey:
+			ecdsaSig := new(struct{ R, S *big.Int })
+			asn1.Unmarshal(sig, ecdsaSig)
+			if !ecdsa.Verify(pub, msg, ecdsaSig.R, ecdsaSig.S) {
+				return errors.New("ecdsa verify failed")
+			}
+		default:
+			return errors.New("unknown public key type")
+		}
+	}
+
+	if pub, ok := priv.Public().(*rsa.PublicKey); ok {
+		var c, m []byte
+		if c, err = rsa.EncryptPKCS1v15(r, pub, ptxt); err != nil {
+			return
+		}
+
+		var decrypter crypto.Decrypter
+		if decrypter, ok = priv.(crypto.Decrypter); !ok {
+			return errors.New("rsa public key but cannot decrypt")
+		}
+
+		if m, err = decrypter.Decrypt(r, c, &rsa.PKCS1v15DecryptOptions{}); err != nil {
+			return
+		}
+		if bytes.Compare(ptxt, m) != 0 {
+			return errors.New("rsa decrypt failed")
+		}
+
+		if m, err = decrypter.Decrypt(r, c, &rsa.PKCS1v15DecryptOptions{SessionKeyLen: len(ptxt)}); err != nil {
+			return
+		}
+		if bytes.Compare(ptxt, m) != 0 {
+			return errors.New("rsa decrypt failed")
+		}
+
+		if m, err = decrypter.Decrypt(r, c, &rsa.PKCS1v15DecryptOptions{SessionKeyLen: len(ptxt) + 1}); err != nil {
+			return
+		}
+		if bytes.Compare(ptxt, m) == 0 {
+			return errors.New("rsa decrypt suceeded despite incorrect SessionKeyLen")
+		}
+	}
+
+	return nil
+}
